requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
pillow==11.0.0
PyJWT==2.9.0
PyMySQL==1.1.1
pytz==2025.2
razorpay==1.4.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
